requirements.txt
streamlit==1.34.0
mysql-connector-python
pandas==2.2.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
openai==1.23.2
